# edge-control-plane/cmd/api/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: edgeCloud Control Plane API
  version: 1.0.0
  description: |
    HTTP API for the edgeCloud managed WebAssembly edge computing platform.
    Supports tenant and worker authentication. Tenants authenticate via
    Bearer API key. Workers authenticate via Bearer JWT (HMAC-SHA256).
servers:
  - url: https://api.edgecloud.dev
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Tenant API key (Bearer token). Obtain via POST /api/v1/keys.
    WorkerJWTAuth:
      type: http
      scheme: bearer
      description: Worker JWT (HMAC-SHA256). Issued to workers by the control plane.
    InternalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: |
        Shared-secret header for trusted service-to-service callers
        (peer control planes pulling artifacts, the edge-ingress
        fetching traffic splits). Token value must match the
        receiving control plane's `internal_token` config. Reject
        requests to `http://` peer URLs — TLS is required to protect
        the token in transit.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code.
              example: NOT_FOUND
            message:
              type: string
              description: Human-readable error detail. Always "internal error" for 5xx.
              example: deployment not found

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    CreateAPIKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: ci-key
        role:
          type: string
          default: developer
          description: "Role for this key: 'developer' or 'owner'."
          example: developer
    CreateAPIKeyResponse:
      type: object
      properties:
        id:
          type: string
          example: k_abc123
        name:
          type: string
          example: ci-key
        role:
          type: string
          example: developer
        token:
          type: string
          description: Raw key shown only once.
          example: ek_...
    APIKeyInfo:
      type: object
      properties:
        id:
          type: string
          example: k_abc123
        name:
          type: string
          example: ci-key
        role:
          type: string
          example: developer
        created_at:
          type: string
          format: date-time
          example: "2026-01-02T03:04:05Z"
    UpdateAPIKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: New name. Omit to leave unchanged.
          example: renamed-key
        role:
          type: string
          description: "New role. Omit to leave unchanged. Valid roles: 'developer', 'owner', 'viewer'."
          example: viewer
...
    Domain:
      type: object
      description: |
        A tenant-owned FQDN bound to a specific app. The status field is
        server-driven: `pending` is the default, `active` is set by a v2
        Caddy webhook on successful ACME issuance, `failed` is set on
        permanent ACME failure. Tenants cannot mutate status directly.
      properties:
        id:
          type: string
          example: dom_3f4a1c9e
        tenant_id:
          type: string
          example: t_acme
        app_name:
          type: string
          example: api
        fqdn:
          type: string
          example: api.acme.com
        status:
          type: string
          enum: [pending, active, failed]
          example: pending
        last_error:
          type: string
          nullable: true
          description: Set by the (v2) Caddy event webhook when ACME fails.
        created_at:
          type: string
          format: date-time
          example: '2026-06-22T12:34:56Z'
        verified_at:
          type: string
          format: date-time
          nullable: true
      required: [id, tenant_id, app_name, fqdn, status, created_at]
    DomainListResponse:
      type: object
      properties:
        domains:
          type: array
          items:
            $ref: '#/components/schemas/Domain'
    AddDomainRequest:
      type: object
      properties:
        fqdn:
          type: string
          description: |
            Fully-qualified domain name. Must be RFC 1035 shape (lowercase,
            ≤253 chars, ≤63 chars per label, no leading/trailing hyphens),
            no wildcards, and not ending in `.edgecloud.dev` (platform-managed
            host).
          example: api.acme.com
      required: [fqdn]
  responses:
    BadRequest:
      description: Bad request — invalid or missing parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: BAD_REQUEST
              message: name is required
    Unauthorized:
      description: Unauthorized — missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: invalid api key
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: deployment not found
...
    QuotaExceeded:
      description: Quota exceeded — too many resources.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: QUOTA_EXCEEDED
              message: max apps quota exceeded
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_ERROR
              message: internal error
  parameters:
    appName:
      name: appName
      in: path
      required: true
      schema:
        type: string
      description: Unique name of the app within the tenant.
paths:
...
  /api/v1/keys:
    post:
      operationId: createAPIKey
      summary: Create a new API key for the authenticated tenant
      description: |
        Issues a new API key scoped to the authenticated tenant.
        The raw key is returned only once — store it securely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
            example:
              name: ci-key
              role: developer
      responses:
        '201':
          description: API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listAPIKeys
      summary: List all API keys for the authenticated tenant
      responses:
        '200':
          description: List of API key metadata (raw keys not included).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/APIKeyInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Auth
  # -------------------------------------------------------------------------
...
  /api/v1/keys/{keyID}:
    put:
      operationId: updateAPIKey
      summary: Update an API key's name or role
      parameters:
        - name: keyID
          in: path
          required: true
          schema:
            type: string
          description: The API key ID to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAPIKeyRequest'
      responses:
        '200':
          description: API key updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKeyInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteAPIKey
      summary: Revoke an API key
      parameters:
        - name: keyID
          in: path
          required: true
          schema:
            type: string
          description: The API key ID to revoke.
      responses:
        '204':
          description: API key revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/keys/{keyID}/rotate:
    post:
      operationId: rotateAPIKey
      summary: Rotate an API key (generate new key, expire old one)
      parameters:
        - name: keyID
          in: path
          required: true
          schema:
            type: string
          description: The API key ID to rotate.
      responses:
        '201':
          description: New API key created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAPIKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Egress allowlist
  # -------------------------------------------------------------------------
...
  /api/v1/apps/{appName}/domains:
    post:
      operationId: addDomain
      summary: Bind a custom FQDN to an existing app
      description: |
        Validates the FQDN shape (RFC 1035, lowercase, ≤253 chars, no
        wildcards, no `.edgecloud.dev` suffix), enforces a per-app quota
        (50 by default), and inserts a row in `pending` state. The ingress
        picks up the new row on its next 30 s poll and Caddy begins ACME
        issuance on the first request to that host.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/appName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDomainRequest'
      responses:
        '201':
          description: Domain row created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: App not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/QuotaExceeded'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: listDomains
      summary: List all custom FQDNs bound to an app
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/appName'
      responses:
        '200':
          description: List of domain rows.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/apps/{appName}/domains/{fqdn}:
    get:
      operationId: getDomain
      summary: Fetch one domain row by FQDN
      description: |
        Returns the row including its `status`, `last_error` (if the v2
        Caddy webhook has populated it), and `verified_at`. Use this to
        diagnose a `failed` row.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/appName'
        - name: fqdn
          in: path
          required: true
          schema:
            type: string
          description: The FQDN to look up.
      responses:
        '200':
          description: Domain row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No row matches (app, fqdn).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: removeDomain
      summary: Unbind a custom FQDN from an app
      description: |
        Deletes the row. The ingress drops the FQDN route on its next
        30 s poll and Caddy's cert for that host is not renewed on expiry.
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/appName'
        - name: fqdn
          in: path
          required: true
          schema:
            type: string
          description: The FQDN to remove.
      responses:
        '204':
          description: Domain row removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No row matches (app, fqdn).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

  # -------------------------------------------------------------------------
  # Internal: Worker-facing
  # -------------------------------------------------------------------------
...